src/zen/pyproject.toml
[tool.poetry]
name = "zen"
version = "0.1.0"
description = ""
authors = ["yuanzx <yuanzx@yunphant.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ib-insync = "^0.9.86"
fastapi = "^0.115.2"
loguru = "^0.7.2"
cachetools = "^5.5.0"
asyncache = "^0.3.1"
types-cachetools = "^5.5.0.20240820"
orjson = "^3.10.9"
jsonrpcserver = "^5.0.9"
sqlmodel = "^0.0.22"
sqlalchemy-stubs = "^0.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
aiosqlite = "^0.20.0"
shelved-cache = "^0.4.0"
click = "^8.1.7"
selenium = "^4.25.0"
pandas-ta = "^0.3.14b0"
colored = "^2.2.4"
numpy = "1.26.4"
moomoo-api = "^8.7.4708"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
